wallet-balance/requirements.txt
moralis
web3
azure-functions
requests
pytz